+++ frontend/src/mod/display/map/renderer/def.d.ts
@@ -6,9 +6,11 @@ export interface ClosureOptions {
 }
 export interface ClosureConfig {
     zLayer: number;
+    tag?: string;
     static?: boolean;
     repeating?: number;
 }
 export type Closure = (ctx: CanvasRenderingContext2D, options: closureOptions) => void;
 
 export type LineData = [Coord, Coord, Color, number, ClosureConfig];
+export type PointData = [Coord, number, Color, ClosureConfig];