+++ frontend/src/pages/dev/editor/testData.ts
@@ -1,15 +1,12 @@
-<script setup lang="ts">
-import { Polygon } from '@/mod/data/polygon/polygon';
-import { Room } from '@/mod/data/room/room';
-import MapDisplay from '@/mod/display/map/MapDisplay.vue';
-import * as graphTools from '@/mod/algorithm/parserTools/graphTools';
+import { Polygon } from '@/mod/data/polygon/polygon'
+import { Room } from '@/mod/data/room/room'
 
 const testPolygon = new Polygon(
     [
         [0, 0],
         [0, 100],
-        [95, 100],
-        [95, 0],
+        [100, 100],
+        [100, 0],
     ],
     [70, 70, 70, 1]
 );
@@ -24,8 +21,8 @@ const testPolygon2 = new Polygon(
 );
 const testPolygon3 = new Polygon(
     [
-        [205, 0],
-        [205, 100],
+        [200, 0],
+        [200, 100],
         [300, 100],
         [300, 0],
     ],
@@ -76,15 +73,12 @@ const testRoom5 = new Room(7102, 1, testPolygon5, 'testRoom5ID');
 const testRoom6 = new Room(7103, 1, testPolygon6, 'testRoom6ID');
 const testRoom7 = new Room(7104, 1, testPolygon7, 'testRoom7ID');
 
-function getRooms() {
-    return [testRoom, testRoom2, testRoom3, testRoom4, testRoom5, testRoom6, testRoom7];
-}
-
-function getGraph(): graphTools.GraphData {
-    return graphTools.generateGraph(getRooms());
-}
-</script>
-
-<template>
-    <MapDisplay ref="display" :gps-coord="[0, 0]" :path-data="getGraph()" :room-data="getRooms()" />
-</template>
+export const testData = [
+	testRoom,
+	testRoom2,
+	testRoom3,
+	testRoom4,
+	testRoom5,
+	testRoom6,
+	testRoom7,
+]