+++ frontend/src/router.ts
@@ -19,16 +19,6 @@ const router = createRouter({
             name: 'dev/editor',
             component: () => import('@/pages/dev/editor/indexEditor.vue'),
         },
-        {
-            path: '/dev/editor/parser',
-            name: 'dev/editor/p',
-            component: () => import('@/pages/dev/editor/indexParser.vue'),
-        },
-        {
-            path: '/dev/editor/parserv2',
-            name: 'dev/editor/p2',
-            component: () => import('@/pages/dev/editor/indexParserv2.vue'),
-        },
     ],
 });